Gemini-Vision/requirements.txt
google-generativeai
streamlit
python-dotenv
pandas
openpyxl